refactor: fix wordpress
