+++ cyberdrop_dl/crawlers/wordpress/bestprettygirl.py
@@ -4,7 +4,7 @@ from typing import ClassVar
 
 from cyberdrop_dl.url_objects import AbsoluteHttpURL
 
-from ._wordpress import WordPressMediaCrawler
+from . import WordPressMediaCrawler
 
 
 class BestPrettyGirlCrawler(WordPressMediaCrawler):
