+++ cyberdrop_dl/crawlers/wordpress/everia.py
@@ -4,7 +4,7 @@ from typing import ClassVar
 
 from cyberdrop_dl.url_objects import AbsoluteHttpURL
 
-from ._wordpress import WordPressHTMLCrawler
+from . import WordPressHTMLCrawler
 
 
 class EveriaClubCrawler(WordPressHTMLCrawler):
